[SC-47216][WARMFIX]Fix "startingTimestamp" behavior for Delta streaming source
## What changes were proposed in this pull request?

Currently when we look up the version for `startingTimestamp`, we will search the latest commit that's **before or at** the timestamp. However, the behavior of `startingTimestamp` should be returning all changes happening **at or after** the timestamp. So we should search the earliest commit that's at or after the timestamp.

Another change is we don't need to require `startingVersion/startingTimestamp` point to a recreatable version. If the json file exists, we should be able to read changes.

## How was this patch tested?

Refactoring the existing tests for `startingVersion` and `startingTimestamp` to cover more edge cases.

## **IMPORTANT** Warmfix instructions

- [X] Other (please describe): Fix an incorrect behavior of a new feature

Author: Shixiong Zhu <zsxwing@gmail.com>

#12333 is resolved by zsxwing/SC-47216.

GitOrigin-RevId: 7918cc10365c95d79f544beb3670f8d7121a79f3
(cherry picked from commit 9fc2acba2bce9869c98f2990dfe57a2667720d1c)