[SC-50564][DELTA][HOTFIX] Do not re-analyze optimized plans while creating Delta tables
There are unfortunately some Catalyst rules in Spark that are not idempotent. These rules can cause data corruption issues if an optimized plan gets re-analyzed. Therefore, we should prevent Delta from having that problem.

In DataSourceV2, the query plan for an AS SELECT statement (such as CREATE TABLE AS SELECT) gets optimized and turned into a SparkPlan. In this phase, for DataSource V1 writers, we pass in a DataFrame created from the optimized plan (and skip re-analysis) as fixed in [Apache Spark](https://github.com/apache/spark/commit/d378dc5f6db6fe37426728bea714f44b94a94861). When Delta was a V1 data source, we didn't have this problem, because we would intercept the analyzed plan, and plug it into a CreateDeltaTableCommand, where the query plan didn't get mucked around with anymore. If we try to go back to the LogicalPlan and to a DataFrame again, the optimized plan can get re-analyzed and cause issues.

This PR is a follow up to  that prevents the re-analysis of an optimized plan during table creation. I the code isn't super clean, but the changes are left to a minimum intentionally so that we can backport this fix.

To the curious: One bug we observed was a manifestation of https://github.com/apache/spark/pull/29805. A literal in the query plan that was propagated during optimization caused the value of the column to completely replace the real value after the plan got re-analyzed and optimized.

A new test suite that checks all DSV2 operations that Delta supports.

If this PR needs to be warmfixed (i.e. merged into the release branch after the code freeze), please follow steps below.

What type of warmfix is this? Please select **exactly one choice**, or write description in Other.

-Regression (e.g. fixing the behavior of a feature that regressed in the current release cycle)
-ES ticket fix (e.g. Customer or internally requested update/fix)
- [ ] Other (please describe):

Make the following updates to this PR:

:warning: *Only* do this if you are targeting the master branch and *additionally* want to have the MergePRSpark job put your commits into additional branches. If your PR already targets the branch you want to warmfix it to, just merge it like a normal PR and do not apply branch labels, milestone, or put "warmfix" in the title before triggering a merge. Track improvements to this confusion at [PLAT-6334](https://databricks.atlassian.net/browse/PLAT-6334).

- [ ] Add `[WARMFIX]` in the title of this PR.
- [ ] Label the PR using label(s) corrsponding to the WARMFIX branch(es). The label name should be in the format of `dbr-branch-a.b` (e.g. `dbr-branch-3.2`), which matches the release branch name for Runtime release `a.b`.
- [ ] Ask your team lead or a staff engineer (or above) to sign off this warmfix and add the `warmfix-approved` label.
- [ ] When merging the PR using the merge script, make sure to get this PR merged into the following branches:

  - The branch against which your PR is opened, and
  - Any extra release branch(es) corresponding to the `dbr-branch-a.b` label(s) applied to your PR.

Author: Burak Yavuz <brkyvz@gmail.com>

GitOrigin-RevId: b3f2793a2cc5ccee7acafb865d145a4b43d54687
(cherry picked from commit 9c3313295dfd7e32df6f95b5aef7ae855d1b84b8)